feat(phase-3): label-free pretrain loader, loop, checkpoints, wandb
Seeds come from ordinary table rows instead of a RelBench task table, with the
SAME sampler as finetuning. One table per batch (entity_table / row_is_root
need it) interleaved by a deterministic round-robin, so every DDP rank picks
the same table at the same step.

Seed tables are weighted by MASKABLE rows, and the payoff is measured: the
task-driven loader yields ZERO seed targets on rel-f1 because the entity table
`drivers` has no maskable column; this one gives every seed a target. Train
seed times are capped at val_timestamp -- the graph is materialized
upto_test_timestamp=False, so without that cap a pretraining batch could
legally contain test-period cells and nothing would notice.

First LR schedule and first disk checkpoints in the repo. Checkpoints split a
portable trunk from per-DB adapters; load_trunk reports missing_adapter and
missing_portable separately, since a trunk never contains adapter keys and
conflating them makes the report useless exactly on a LODO target.

Portability verified, not assumed: rel-f1 (9 tables/13 roles/45 cols) vs
rel-trial (15/15/103) share zero differently-shaped keys, and a scan for
learned tensors dimensioned by C/n_tables/K finds 0 hits outside
cell_encoders. A rel-f1 trunk loads onto rel-trial and runs.

Two bugs found: split_state_dict used startswith() against LightningModule
keys prefixed `model.`, so every adapter tensor was classified portable; and
under autocast the sparse path's index_add raises where dense `+` promotes.

GATE: rel-f1 300 steps, flex+sparse+8/2 experts, val/loss 0.170. 320 tests pass.

+++ gloss/model/column_encoder.py
@@ -149,9 +149,13 @@ class CellEncoder(nn.Module):
             b_idx, s_idx, row_idx, col_idx = cb.cell_placement[nt]
             b_idx, s_idx = b_idx.to(dev), s_idx.to(dev)
             row_idx, col_idx = row_idx.to(dev), col_idx.to(dev)
-            h[b_idx, s_idx] = cell[row_idx, col_idx]
+            # `.to(h.dtype)`: under `precision='bf16-mixed'` the stype encoders return bf16 while `h`
+            # is the fp32 accumulator, and an index_put across dtypes raises rather than promoting.
+            # Keeping the scatter buffer in fp32 is also the right call numerically — autocast recasts
+            # at the next matmul anyway, so nothing is gained by narrowing it here.
+            h[b_idx, s_idx] = cell[row_idx, col_idx].to(h.dtype)
         if cell_mask is not None:
             cm = cell_mask.to(dev).unsqueeze(-1)
-            h = torch.where(cm, self.masked_token(cb), h)
+            h = torch.where(cm, self.masked_token(cb).to(h.dtype), h)
         mask = (~cb.is_padding).to(dev).unsqueeze(-1)
         return h * mask
